JI-3997 use webpack
