+++ package.json
@@ -4,20 +4,22 @@
   "description": "Provides H5P with the popular, easy to use and lightweight 3D library: three.js",
   "main": "builder.js",
   "scripts": {
-    "build": "node builder.js",
+    "build": "webpack",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/h5p/h5p-three-js.git"
   },
-  "author": "three.js authors",
+  "author": "three.js authors & H5P Group",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/h5p/h5p-three-js/issues"
   },
   "homepage": "https://github.com/h5p/h5p-three-js#readme",
-  "dependencies": {
-    "three": "0.101.0"
+  "devDependencies": {
+    "three": "0.101.0",
+    "webpack": "^5.88.2",
+    "webpack-cli": "^5.1.4"
   }
 }
